Magisk: don't ship Magiskmanager.
It' not a good idea to have MagiskManager installed as
system app. Leads to all kind of issues, if you try to
use beta versions or if you want to uninstall Magisk
all together. So we remove it again.

The MagiskManager stub is installed by the flashable
magisk.zip, anyway. On reboot the user is asked to
update MagiskManager from playstore. So this is no big
deal.

Remember to wipe system and reinstall gapps or microg.

Change-Id: I12e8a89c7f577b3c6b82cbde81f78b4d89b156e4